XML and CSV configuration loaders

+++ src/Avans.FlatGalaxy.Models/CelestialBodies/CelestialBody.cs
@@ -9,9 +9,9 @@ namespace Avans.FlatGalaxy.Models.CelestialBodies
 
         public int Y { get; set; }
 
-        public int VX { get; set; }
+        public double VX { get; set; }
 
-        public int VY { get; set; }
+        public double VY { get; set; }
 
         public int Radius { get; set; }
 
@@ -19,7 +19,7 @@ namespace Avans.FlatGalaxy.Models.CelestialBodies
 
         public ICollisionState CollisionState { get; set; }
 
-        public CelestialBody(int x, int y, int vx, int vy, int radius, Color color, ICollisionState collisionState)
+        public CelestialBody(int x, int y, double vx, double vy, int radius, Color color, ICollisionState collisionState)
         {
             X = x;
             Y = y;
